Add support for Netlify
Instead of deploying to S3, switch to deploying to Netlify, as it
supports HTTPS.

One thing to note about Netlify is that it's kind of weird about
caching. It actually doesn't cache *anything* ([1]), choosing to rely on
the speed on their CDN. However, we want to cache font files so that we
don't get a FOUC when each page loads. In any case, Middleman will
append a hash to each filename, so if a file gets updated it'll bust the
cache. So we use netlify.toml to override the Cache-Control header to
what we want.

[1]: https://www.netlify.com/blog/2017/02/23/better-living-through-caching/