Revert "Add support for Netlify"
This reverts commit 267acdd3f5cb3c580507cc435f266bf2b2dcb3d2.

Netlify doesn't work at all. It's definitely fast but there's one fatal
flaw: it doesn't cache anything. They claim this makes things easier to
manage on their end, and maybe it does, except that I'm fingerprinting
everything and so there's no need to turn off caching. It causes FOUC
and that's something we want to avoid especially on a static site.
Supposedly there's a way you can override the Cache-Control headers to
get around this but that wasn't working. I just figured out a way to use
Amazon CloudFront and Certificate Manager so I'm going to try that
instead.